requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy>=2.0.0,<2.1.0
asyncpg>=0.28.0
alembic>=1.12.0

# Supabase integration
supabase==2.0.2
postgrest==0.13.0

# Redis and caching
redis>=5.0.0
# aioredis>=2.0.1  # Temporarily disabled due to distutils issue

# Background jobs
celery==5.3.4
celery[redis]==5.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# External API integrations
httpx>=0.24.0,<0.25.0
google-generativeai==0.3.2
sendgrid==6.10.0
tweepy==4.14.0

# Data validation and serialization
pydantic>=2.4.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
email-validator>=2.0.0

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
validators==0.22.0
feedparser>=6.0.11,<7.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Production
gunicorn==21.2.0
greenlet>=2.0.0